# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GalaxyBoards API",
    "description": "An ASP.NET Core Web API for creating and managing Kanban-style boards",
    "version": "v1"
  },
  "paths": {
...
    "/api/Tickets": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Queries stored tickets",
        "description": "All parameters are optional. A request without any parameters set will return all stored tickets without any filtering or sorting applied.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filters tickets by those whose names or descriptions contain 1 or more of the provided search terms/words",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "The ticket property to sort tickets by, before filtering",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "orderAscending",
            "in": "query",
            "description": "The sort direction to use when `orderBy` is provided. Set to `true` for ascending order. Defaults to descending order.",
            "style": "form",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "The maximum tickets that will be returned. May be used for pagination when combined with `offset`.",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "The tickets to skip. May be used for pagination when combined with `limit`.",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "projectId",
            "in": "query",
            "description": "Filters tickets by their associated project ID",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketBriefQueryResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tickets"
        ],
        "summary": "Creates a new ticket",
        "requestBody": {
          "description": "Ticket data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketPostData"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketPostData"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TicketPostData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully created ticket",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/Tickets/{id}": {
      "get": {
        "tags": [
          "Tickets"
        ],
        "summary": "Looks up a ticket by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of ticket to retrieve",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "404": {
            "description": "Ticket not found"
          }
        }
      },
      "put": {
        "tags": [
          "Tickets"
        ],
        "summary": "Updates an existing ticket",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of ticket to update",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Ticket update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TicketUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket successfully updated"
          },
          "400": {
            "description": "Invalid payload data"
          },
          "404": {
            "description": "Ticket not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Tickets"
        ],
        "summary": "Deletes a ticket by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the ticket to delete",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket successfully deleted"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PostResult": {
        "required": [
          "createdId"
        ],
        "type": "object",
        "properties": {
          "createdId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Data returned in response to a successful POST request. Contains the ID of the newly created entity."
      },
...
      "TicketBrief": {
        "required": [
          "id",
          "links",
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "minLength": 1,
            "type": "string"
          },
          "links": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false,
        "description": "Describes a Ticket entity in brief"
      },
      "TicketBriefQueryResult": {
        "required": [
          "items",
          "totalRecords"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketBrief"
            }
          },
          "totalRecords": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Data returned in response to a GET request to queryable endpoints. Contains a list of records matching any given query parameters, and the total count."
      },
      "TicketDetail": {
        "required": [
          "description",
          "id",
          "links",
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "minLength": 1,
            "type": "string"
          },
          "links": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "description": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "Describes a Ticket entity in full"
      },
...
      "TicketPostData": {
        "required": [
          "description",
          "name",
          "projectId",
          "status"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Required data for Ticket creation"
      },
      "TicketUpdate": {
        "required": [
          "description",
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "status": {
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false,
        "description": "Ticket update data"
      }
    }
  }
}